refactor(search): split vector-search.repository and search-engine (#611)
Decompose two large search orchestrators into composition sub-modules
while preserving public import paths and FTS5/reflection_notes behavior.

Closes #611
